requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
slack_sdk
